canRdtTypeEqual call in CRDTs should use explicit type parameter
Calling callRdtTypeEqual can automatically bind the implict TypeTag that
we usually have in CRDTs but it can be confusing (as to what type gets
used) and fail to compile in case of multiple implicit TypeTag
instances. Using explicit tpe parameter on the call-stie makes it much
more understandable.

Also renamed type parameter in most CRDT from T to E, since they are
container types and E denotes better the intention behind it, i.e.
`E`lement.

Added some explanation and docs to RdtTypeEquals. The RdtTypeEquals
should use self-annotation for RDT, since it is clearly to be used just
with RDTs.